requirements.txt
streamlit==1.33.0
langchain==0.1.14
langchain-google-genai==0.0.11
google-generativeai==0.3.2
faiss-cpu==1.7.4
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
pydantic==2.6.4
tenacity==8.2.3
pytest==8.1.1
matplotlib==3.8.3
